# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Second Brain API",
    "description": "Knowledge management and learning system API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/review/cards": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Create Card",
        "description": "Create a new spaced repetition card.\n\nCard is created in NEW state with FSRS initial parameters.",
        "operationId": "create_card_api_review_cards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/cards/{card_id}": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Card",
        "description": "Get a card by ID.",
        "operationId": "get_card_api_review_cards__card_id__get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/due": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Due Cards",
        "description": "Get cards due for review.\n\nReturns cards ordered by due date (oldest first) with a review forecast.",
        "operationId": "get_due_cards_api_review_due_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Maximum cards to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum cards to return"
          },
          {
            "name": "topic",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by topic tag",
              "title": "Topic"
            },
            "description": "Filter by topic tag"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DueCardsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/rate": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Rate Card",
        "description": "Submit a review rating for a card.\n\nFSRS ratings:\n- AGAIN (1): Complete failure, reset to learning\n- HARD (2): Significant difficulty, shorter interval\n- GOOD (3): Correct with effort, normal interval\n- EASY (4): Too easy, longer interval\n\nReturns the new scheduling information.",
        "operationId": "rate_card_api_review_rate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/evaluate": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Evaluate Card Answer",
        "description": "Evaluate a typed answer for a card using LLM.\n\nThis enables \"active recall\" mode where users type their answer\ninstead of just flipping the card. The LLM evaluates the answer\nsemantically and returns an appropriate FSRS rating.\n\nUse the returned rating with the /rate endpoint to save the review.\n\nReturns:\n- rating: FSRS rating (1-4) based on answer quality\n- is_correct: True if rating >= 3 (Good or Easy)\n- feedback: Explanation of what was correct/incorrect\n- expected_answer: The correct answer for comparison",
        "operationId": "evaluate_card_answer_api_review_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardEvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardEvaluateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/stats": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Card Stats",
        "description": "Get card statistics.\n\nReturns counts by state, average stability/difficulty, and due counts.",
        "operationId": "get_card_stats_api_review_stats_get",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by topic tag",
              "title": "Topic"
            },
            "description": "Filter by topic tag"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/generate": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Generate Cards",
        "description": "Generate spaced repetition cards for a topic on-demand.\n\nUses existing content and LLM to generate flashcards for the specified topic.\nUseful when starting a review session for a topic with few or no cards.",
        "operationId": "generate_cards_api_review_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardGenerationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/ensure-cards": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Ensure Cards For Topic",
        "description": "Ensure a minimum number of cards exist for a topic.\n\nIf fewer than `minimum` cards exist, generates more using LLM.\nReturns immediately if enough cards already exist.",
        "operationId": "ensure_cards_for_topic_api_review_ensure_cards_post",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Topic path",
              "title": "Topic"
            },
            "description": "Topic path"
          },
          {
            "name": "minimum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Minimum cards required",
              "default": 5,
              "title": "Minimum"
            },
            "description": "Minimum cards required"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardGenerationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardCreate": {
        "properties": {
          "card_type": {
            "type": "string",
            "title": "Card Type",
            "description": "Card type: concept, fact, application, cloze, code"
          },
          "front": {
            "type": "string",
            "title": "Front",
            "description": "Front side (question/prompt)"
          },
          "back": {
            "type": "string",
            "title": "Back",
            "description": "Back side (answer)"
          },
          "hints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hints",
            "description": "Progressive hints"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Topic tags"
          },
          "content_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Id",
            "description": "Source content ID"
          },
          "concept_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Id",
            "description": "Related concept ID in Neo4j"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "Programming language"
          },
          "starter_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starter Code",
            "description": "Initial code template"
          },
          "solution_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solution Code",
            "description": "Reference solution"
          },
          "test_cases": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Cases",
            "description": "Test cases for code cards"
          }
        },
        "type": "object",
        "required": [
          "card_type",
          "front",
          "back"
        ],
        "title": "CardCreate",
        "description": "Request to create a new spaced repetition card.\n\nExtends CardBase with optional source linking to content and Neo4j concepts.\nFor code cards, includes language, starter template, solution, and test cases\nthat enable automated evaluation in the code sandbox."
      },
      "CardEvaluateRequest": {
        "properties": {
          "card_id": {
            "type": "integer",
            "title": "Card Id",
            "description": "Card ID"
          },
          "user_answer": {
            "type": "string",
            "title": "User Answer",
            "description": "User's typed answer"
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "user_answer"
        ],
        "title": "CardEvaluateRequest",
        "description": "Request to evaluate a typed answer for a card."
      },
      "CardEvaluateResponse": {
        "properties": {
          "card_id": {
            "type": "integer",
            "title": "Card Id"
          },
          "rating": {
            "type": "integer",
            "maximum": 4,
            "minimum": 1,
            "title": "Rating",
            "description": "FSRS rating (1-4)"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct",
            "description": "Whether the answer is considered correct"
          },
          "feedback": {
            "type": "string",
            "title": "Feedback",
            "description": "Explanation for the learner"
          },
          "key_points_covered": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Points Covered"
          },
          "key_points_missed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Points Missed"
          },
          "expected_answer": {
            "type": "string",
            "title": "Expected Answer",
            "description": "The correct answer for comparison"
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "rating",
          "is_correct",
          "feedback",
          "expected_answer"
        ],
        "title": "CardEvaluateResponse",
        "description": "Response from card answer evaluation."
      },
...
      "CardResponse": {
        "properties": {
          "card_type": {
            "type": "string",
            "title": "Card Type",
            "description": "Card type: concept, fact, application, cloze, code"
          },
          "front": {
            "type": "string",
            "title": "Front",
            "description": "Front side (question/prompt)"
          },
          "back": {
            "type": "string",
            "title": "Back",
            "description": "Back side (answer)"
          },
          "hints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hints",
            "description": "Progressive hints"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Topic tags"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "content_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Id"
          },
          "concept_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Id"
          },
          "state": {
            "$ref": "#/components/schemas/CardState",
            "default": "new"
          },
          "stability": {
            "type": "number",
            "title": "Stability",
            "description": "Memory stability in days",
            "default": 0
          },
          "difficulty": {
            "type": "number",
            "minimum": 0,
            "title": "Difficulty",
            "description": "FSRS card difficulty (can exceed 1.0 after reviews)",
            "default": 0.3
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "title": "Due Date"
          },
          "last_reviewed": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reviewed"
          },
          "repetitions": {
            "type": "integer",
            "title": "Repetitions",
            "default": 0
          },
          "lapses": {
            "type": "integer",
            "title": "Lapses",
            "default": 0
          },
          "total_reviews": {
            "type": "integer",
            "title": "Total Reviews",
            "default": 0
          },
          "correct_reviews": {
            "type": "integer",
            "title": "Correct Reviews",
            "default": 0
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "card_type",
          "front",
          "back",
          "id",
          "due_date"
        ],
        "title": "CardResponse",
        "description": "Card response with FSRS state.\n\nIncludes the full FSRS scheduling state: stability (memory strength in days),\ndifficulty (0-1 scale), and current learning state. The due_date indicates when\nthe card should next be reviewed based on the FSRS-4.5 optimal retention algorithm."
      },
      "CardReviewRequest": {
        "properties": {
          "card_id": {
            "type": "integer",
            "title": "Card Id",
            "description": "Card ID to review"
          },
          "rating": {
            "$ref": "#/components/schemas/Rating",
            "description": "Self-assessment rating (1-4)"
          },
          "response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response",
            "description": "Optional response text"
          },
          "time_spent_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Spent Seconds",
            "description": "Time spent on review"
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "rating"
        ],
        "title": "CardReviewRequest",
        "description": "Request to submit a card review rating.\n\nThe rating follows FSRS conventions: Again (1) for forgotten, Hard (2) for difficult\nrecall, Good (3) for successful recall with effort, Easy (4) for effortless recall.\nTime spent is tracked for analytics and can inform future difficulty adjustments."
      },
      "CardReviewResponse": {
        "properties": {
          "card_id": {
            "type": "integer",
            "title": "Card Id"
          },
          "new_state": {
            "$ref": "#/components/schemas/CardState"
          },
          "new_stability": {
            "type": "number",
            "title": "New Stability"
          },
          "new_difficulty": {
            "type": "number",
            "title": "New Difficulty"
          },
          "next_due_date": {
            "type": "string",
            "format": "date-time",
            "title": "Next Due Date"
          },
          "scheduled_days": {
            "type": "integer",
            "title": "Scheduled Days",
            "description": "Days until next review"
          },
          "was_correct": {
            "type": "boolean",
            "title": "Was Correct",
            "description": "Whether rating indicates success"
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "new_state",
          "new_stability",
          "new_difficulty",
          "next_due_date",
          "scheduled_days",
          "was_correct"
        ],
        "title": "CardReviewResponse",
        "description": "Response after submitting a card review.\n\nContains the updated FSRS state after processing the review. The scheduled_days\nfield shows when the card will next be due, which can range from minutes (for\nAgain ratings) to months (for well-learned Easy cards with high stability)."
      },
...
      "CardStats": {
        "properties": {
          "total_cards": {
            "type": "integer",
            "title": "Total Cards"
          },
          "cards_by_state": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Cards By State",
            "description": "Count per CardState"
          },
          "avg_stability": {
            "type": "number",
            "title": "Avg Stability",
            "default": 0
          },
          "avg_difficulty": {
            "type": "number",
            "title": "Avg Difficulty",
            "default": 0
          },
          "due_today": {
            "type": "integer",
            "title": "Due Today",
            "default": 0
          },
          "overdue": {
            "type": "integer",
            "title": "Overdue",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "total_cards"
        ],
        "title": "CardStats",
        "description": "Statistics about cards in the system.\n\nProvides aggregate metrics for the card collection including state distribution,\naverage stability/difficulty, and review workload. Useful for understanding\noverall system health and planning study capacity."
      },
...
      "DueCardsResponse": {
        "properties": {
          "cards": {
            "items": {
              "$ref": "#/components/schemas/CardResponse"
            },
            "type": "array",
            "title": "Cards"
          },
          "total_due": {
            "type": "integer",
            "title": "Total Due"
          },
          "review_forecast": {
            "$ref": "#/components/schemas/ReviewForecast"
          }
        },
        "type": "object",
        "required": [
          "cards",
          "total_due",
          "review_forecast"
        ],
        "title": "DueCardsResponse",
        "description": "Response with due cards and forecast.\n\nReturns cards ordered by due date (oldest first) along with the total count\nof due cards and a forecast of upcoming reviews. The forecast helps users\nunderstand their review backlog and plan study sessions accordingly."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}